Refactor dungeon generation and move player/prefabs for testing purposes
Refactored Room and Dungeon classes into NCL::CSC8508::DungeonGen namespace.
